Multiple draw calls implemented for batch rendering.

+++ Hazel/src/Hazel/Rendering/Textures/Texture.h
@@ -9,6 +9,6 @@ namespace Hazel
 
 		virtual size_t GetWidth() const = 0;
 		virtual void SetData(const void *data) = 0;
-		virtual void Bind(unsigned int slot = 0) const = 0;
+		virtual void Bind(size_t slot = 0) const = 0;
 	};
 }
